Fixing the behaviour of removeRows.
There was a bug when more than one table view were displayed : the
shortcut was ambiguous because it was declared for each table view.
The emitted signal was "activatedAmbiguously" and it was emitted for
one of the table view, but not necessarily the one that had the focus.

I fixed it by declaring only one shortcut, and by finding out wich table
view has the focus.