Fixing the undo mechanisms on indexes edition.
There was too many history events started for one edition. First problem
was that the start_history_event method was called in setData, wich was
itself called each time QGitDelegate.setModelData was called (once for
every index). The second problem was that in commitAndCloseEditor, we
were sending twice a signal that would call setModelData (commit and
close).